+++ aui.core/src/AUI/IO/MultipleInputStream.h
@@ -40,7 +40,7 @@ public:
 
     ~MultipleInputStream() override = default;
 
-    int read(char* dst, int size) override;
+    size_t read(char* dst, size_t size) override;
 };
