+++ aui.core/src/AUI/IO/StringStream.h
@@ -37,6 +37,6 @@ public:
 
 	void seekRead(size_t position);
 
-	int read(char* dst, int size) override;
-    int write(const char *src, int size) override;
+	size_t read(char* dst, size_t size) override;
+    size_t write(const char *src, size_t size) override;
 };
