+++ aui.curl/src/AUI/Curl/ACurl.h
@@ -74,7 +74,7 @@ public:
 	explicit ACurl(Builder&& builder);
 	~ACurl();
 
-	int read(char* dst, int size) override;
+	size_t read(char* dst, size_t size) override;
 
 	int64_t getContentLength() const;
 };
